implement more difference navigation
i did not really like the old way of skipping half a page and then
searching for a difference, instead just jump to the next run of
differents.

also implement jumping up bceause that's useful and was not implemented
because i pondered over keybindings (shift+space is not a Thing in
terminals). in the end i left space for ui-backcomp but the primary way
to do this should be shift+up/down and ctrl+up/down (it feels more natural
to me at least).