remove effects before scroll
on some terminals (like cmd.exe or the freebsd boot terminal),
new cells that come into the screen from scrolling have the effect
that is last applied. for biodiff, this is always inverted since the
top/bottom bar are drawn last, resulting in the very right side
of the window getting painted white. we prevent this by clearing
the effect before scrolling